plugin-table: directly use TableConfig
